refactor: extract grouping matches by day into helper functions
